requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.8.0rc2
Django==3.2.25
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.10
idna==3.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
Pillow==9.5.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.15.0
